Add files via upload
This Python project serves as a basic shopping bill calculator to practice fundamental programming concepts. It allows users to input the price of items and calculates the total bill. If the user inputs 'q', it finalizes the bill, provides change if necessary, and ends the transaction. This project is ideal for beginners looking to strengthen their Python skills and understanding of basic arithmetic operations.